form input same colors, new server build, reorganized server auth files
