+++ app/src/components/album/list/update-image-form/update-image-form.html
@@ -5,13 +5,13 @@
   </div>
   <!-- Form -->
   <form ng-submit="$ctrl.submit()">
-    <md-input-container>
+    <md-input-container class="md-accent">
       <input placeholder="Title" ng-model="$ctrl.image.title">
     </md-input-container>
-    <md-input-container>
+    <md-input-container class="md-accent">
       <input placeholder="Description" ng-model="$ctrl.image.description">
     </md-input-container>
-    <md-input-container>
+    <md-input-container class="md-accent">
       <input placeholder="Image URL" ng-model="$ctrl.image.link">
     </md-input-container>
     <div>
