+++ app/src/components/list-albums/update-album-form/update-album-form.html
@@ -5,7 +5,7 @@
   </div>
   <!-- Form -->
   <form ng-submit="$ctrl.submit()">
-    <md-input-container>
+    <md-input-container class="md-accent">
       <input placeholder="Title" ng-model="$ctrl.album.title">
     </md-input-container>
     <md-button type="submit">Send</md-button>
