+++ server/auth/ensureAuth.js
