+++ server/auth/token.js
