--- server/lib/public/index.html
@@ -1,11 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-	<meta charset="UTF-8">
-	<meta name="viewport" content="width=device-width, initial-scale=1.0">
-	<title>Image Gallery</title>
-<link href="/styles/bundle.css" rel="stylesheet"></head>
-<body>
-		<app></app>
-<script type="text/javascript" src="bundle.js"></script></body>
-</html>
